==> script_5.Rmd <==
---
title: "Text mining methods"
author: "bia"
date: "15/01/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(magrittr, include.only = "%>%")

dataset_corpus_raw <- readr::read_rds("data/dataset_corpus_raw.Rds")

head(dataset_corpus_raw)
```

```{r}
dataset_corpus_filtered <- dataset_corpus_raw %>%
  dplyr::mutate(ano = lubridate::year(data_reuniao)) %>%
  dplyr::mutate(
    crise = dplyr::case_when(
      ano %in% c(2008:2012) ~  "antes_crise" ,
      ano %in% c(2013:2015) ~  "durante_crise" ,
      ano %in% c(2016:2020) ~  "pos_crise" ,
      TRUE ~ "outros"
      
    )
  ) %>% 
  dplyr::filter(crise != "outros")


dataset_corpus_filtered  %>% readr::write_rds("data/dataset_corpus_filtered.Rds")


```



```{r}
united_clean_text <- dataset_corpus_filtered %>%
  dplyr::mutate(
    texto = stringr::str_replace_all(text, "_", " "),
    texto = stringr::str_squish(texto)
    
  )

head(united_clean_text)

small_dataset <-   united_clean_text %>%
  dplyr::select(comite, ano, crise, texto) 


head(small_dataset)

# deletando arquivos intermediarios para nao ficar pesado
rm(dataset_corpus_raw)
rm(united_clean_text)
rm(dataset_corpus_filtered)
```




## Bigrams

```{r}
tokens_bigram <- small_dataset %>%
  tidytext::unnest_tokens(bigram, texto, token = "ngrams", n = 2)


# deletando arquivos intermediarios para nao ficar pesado
rm(small_dataset)
```


```{r}
bigrams_separated <- tokens_bigram %>%
  tidyr::separate(bigram, c("word1", "word2"), sep = " ")

# deletando arquivos intermediarios para nao ficar pesado
rm(tokens_bigram)


source("stop_words.R", encoding = "UTF-8")


bigrams_filtered_united <- bigrams_separated %>%
  dplyr::filter(!word1 %in% stopwords_br$word) %>%
  dplyr::filter(!word2 %in% stopwords_br$word) %>%
  dplyr::mutate(
    word1 = stringr::str_replace_all(word1, pattern = "[0-9]+", replacement = NA_character_),
    
    word2 = stringr::str_replace_all(word2, pattern = "[0-9]+", replacement = NA_character_)
  ) %>%
  tidyr::drop_na() %>%
  tidyr::unite(bigram, word1, word2, sep = " ") %>%
  dplyr::filter(!bigram %in% stopwords_br_bigram$word) %>%
  dplyr::mutate(
    bigram = dplyr::case_when(
      bigram == "câmara técnica" ~ "câmara(s) técnica(s)",
      bigram == "câmaras técnicas" ~ "câmara(s) técnica(s)",
      TRUE ~ bigram
    )
  )



# new bigram counts:
bigram_counts <- bigrams_filtered_united  %>%
  dplyr::group_by(crise, comite) %>%
  dplyr::count(bigram, sort = TRUE)

# bigram_counts

```

# Export bigram counts

```{r}
bigram_counts %>% readr::write_rds("data/bigram_counts.Rds")
```
